Validation Errors -> Conflict 409

+++ src/Application/Artists/Commands/DeleteArtistCommand.cs
@@ -1,7 +1,9 @@
+using ChinookStore.Application._Common.Exceptions;
 using ChinookStore.Application._Common.Extensions;
 using ChinookStore.Application._Common.Interfaces;
 using ChinookStore.Domain.Entities;
 using MediatR;
+using Microsoft.EntityFrameworkCore;
 
 namespace ChinookStore.Application.Artists.Commands;
 
@@ -12,6 +14,12 @@ public class DeleteArtistCommandHandler(IRepository repository) : IRequestHandle
     public async Task Handle(DeleteArtistCommand request, CancellationToken cancellationToken)
     {
         var artist = await repository.Query<Artist>().FirstByIdAsync(request.Id, cancellationToken);
+
+        if (await repository.Query<Album>().AnyAsync(a => a.Artist.Id == request.Id, cancellationToken))
+        {
+            throw new ValidationErrorsException("There are one or more albums related to this artist.");
+        }
+
         repository.Delete(artist);
 
         await repository.SaveChangesAsync(cancellationToken);
